requirements.txt
streamlit
PyPDF2
faiss-cpu
sentence-transformers